requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
colored==1.4.3
Django==4.0.4
idna==3.3
numpy==1.22.3
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
typeguard==2.13.3
urllib3==1.26.9
whitenoise==6.1.0
